-Resolve issue of Full Height Drawer. -Resolve issue of click of Menu open Slide Menu.
