+++ src/Components/SplashScreenScene.js
@@ -1,85 +1,77 @@
-import React, { Component } from 'react';
-import {
-	Platform,
-	StyleSheet,
-	Text,
-	View,
-	Image,
-	TouchableOpacity,
-	TouchableHighlight,
-	RefreshControl
-} from 'react-native';
-import { connect } from 'react-redux';
-import { setTopNews } from '../redux/actions/TopHeadingNewsActions';
-import Pulse from 'react-native-pulse';
-import GridView from 'react-native-gridview';
-import HamburgerIcon from './HamburgerIcon';
+import React, { PureComponent } from 'react';
+import { Platform, StyleSheet, Text, View, Image, TouchableOpacity } from 'react-native';
+import NewsDetailScene from './NewsDetailScene';
+import NewsDetail from './NewsDetail';
+import NewsScene from './NewsScene';
+import { createMaterialTopTabNavigator, createStackNavigator } from 'react-navigation';
 
-class SplashScreenScene extends Component {
-	constructor(props) {
-		super(props);
-		this.state = {
-			refreshing: false
-		};
-	}
-
-	componentDidMount() {
-		this.props.setTopNews();
-	}
-
-	_onRefresh() {
-		if (this.state.refreshing) {
-			return;
-		}
-		this.setState({ refreshing: true });
-		this.props.setTopNews();
-		if (this.props.showApiData) {
-			this.setState({ refreshing: false });
+const isTabBarVisible = routes => {
+	var isVisible = true;
+	const screenNames = ['NewsDetailScene'];
+	if (routes.length > 0) {
+		for (let i = 0; i < routes.length; i++) {
+			if (screenNames[0] == routes[i].routeName) {
+				isVisible = false;
+			}
 		}
 	}
-
+	return isVisible;
+};
+const AppTabBar = createMaterialTopTabNavigator(
+	{
+		tab1: { screen: NewsScene },
+		tab2: { screen: NewsDetail }
+	},
+	{
+		navigationOptions: ({ navigation }) => ({
+			//navigation.state.routeName
+			tabBarVisible: navigation.state.routes == undefined ? true : isTabBarVisible(navigation.state.routes),
+			header: null,
+			tabBarOptions: {
+				activeTintColor: 'rgb(255,87,34)',
+				inactiveTintColor: 'white',
+				upperCaseLabel: true,
+				style: {
+					backgroundColor: 'rgb(2,136,141)'
+				},
+				labelStyle: {
+					fontSize: 15,
+					fontStyle: 'bold'
+				},
+				indicatorStyle: {
+					backgroundColor: 'rgb(255,87,34)',
+					height: 3
+				}
+			}
+		})
+	}
+);
+class SplashScreenScene extends PureComponent {
+	static router = AppTabBar.router;
 	render() {
-		const self = this;
-		const { showApiData, articles } = this.props;
-		console.log(showApiData);
 		return (
 			<View style={styles.container}>
-				{!showApiData && <Pulse color="rgb(2,136,141)" numPulses={5} diameter={100} speed={10} duration={2000} />}
-				{showApiData && (
-					<View>
-						<GridView
-							data={articles}
-							itemsPerRow={2}
-							refreshControl={
-								<RefreshControl refreshing={this.state.refreshing} onRefresh={this._onRefresh.bind(this)} />
-							}
-							renderItem={(item, sectionID, rowID, itemIndex, itemID) => {
-								console.log(item, ' ' + sectionID, ' ' + rowID, ' ' + itemIndex, ' ' + itemID);
-								return (
-									<TouchableOpacity
-										onPress={() => {
-											this.props.navigation.navigate('NewsDetailScene', { index: itemID });
-										}}
-									>
-										<View style={{ backgroundColor: 'white', margin: 7, height: 200, borderRadius: 8 }}>
-											<Image
-												source={{ uri: item.urlToImage }}
-												style={{
-													width: '100%',
-													height: 120,
-													resizeMode: 'cover',
-													borderTopLeftRadius: 10,
-													borderTopRightRadius: 10
-												}}
-											/>
-											<Text style={{ marginTop: 5, padding: 5 }}>{item.title}</Text>
-										</View>
-									</TouchableOpacity>
-								);
-							}}
+				<View
+					style={{
+						height: 50,
+						backgroundColor: 'rgb(2,136,141)',
+						flexDirection: 'row',
+						alignItems: 'center'
+					}}
+				>
+					<TouchableOpacity
+						onPress={() => {
+							this.props.navigation.openDrawer();
+						}}
+					>
+						<Image
+							source={{ uri: 'https://reactnativecode.com/wp-content/uploads/2018/04/hamburger_icon.png' }}
+							style={{ width: 25, height: 25, marginLeft: 5 }}
 						/>
-					</View>
-				)}
+					</TouchableOpacity>
+					<Text style={{ textSize: 15, color: 'white', marginLeft: 20 }}>Samachar</Text>
+				</View>
+				<AppTabBar navigation={this.props.navigation} />
 			</View>
 		);
 	}
@@ -91,19 +83,4 @@ const styles = StyleSheet.create({
 	}
 });
 
-const mapDispatchToProps = dispatch => {
-	return {
-		setTopNews: props => {
-			dispatch(setTopNews(props));
-		}
-	};
-};
-
-const mapStateToProps = state => {
-	return {
-		articles: state.new_reducer.articles,
-		showApiData: state.new_reducer.showApiData
-	};
-};
-
-export default connect(mapStateToProps, mapDispatchToProps)(SplashScreenScene);
+export default SplashScreenScene;
